Add results with graphs for each category (question)
Add a whole lot of dependencies (underscore.js, d3.js) to generate
results for each question (show which country the answer resembles
the most, the minimum and maximum values, and Iceland... because
yay! Iceland).

This section (the categories) will have to be hidden in the final
webapp and shown only after answering all questions but for dev
purposes it's good to show it on the page.