Return false from click function to avoid hashtag in url
When clicking on the <a href="#questions"> on the welcome page
the url changed even if we just want to scroll and the scrollTo
feature has this but when splitting it up into a reusable function
the return false statement went with it so the click changed the
url. Minor bug fix.